Add constraint like_memo & like_comment, Remove like toggle(RESTful)

+++ MemoSquare/views_comment.py
@@ -1,9 +1,11 @@
+from django.shortcuts import get_object_or_404
 from rest_framework.decorators import permission_classes, api_view, renderer_classes
 from rest_framework.response import Response
 from rest_framework import permissions, status
 from MemoSquare.models import Comment, LikeComment
 from rest_framework.renderers import JSONRenderer
 from MemoSquare.serializers import CommentSerializer
+from django.db.utils import IntegrityError
 
 
 @api_view(['GET', 'POST'])
@@ -53,15 +55,23 @@ def update_delete(request, pk):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-@api_view(['POST'])
+@api_view(['POST', 'DELETE'])
 @permission_classes((permissions.IsAuthenticated,))
 def like_unlike(request, pk):
-    try:
-        like = LikeComment.objects.get(user=request.user, comment_id=pk)
+    comment = get_object_or_404(Comment, pk=pk)
+
+    if request.method == 'POST':
+        try:
+            LikeComment.objects.create(user=request.user, comment=comment)
+        except IntegrityError:
+            pass
+        result = 'liked'
+    elif request.method == 'DELETE':
+        try:
+            like = LikeComment.objects.get(user=request.user, comment=comment)
+        except LikeComment.DoesNotExist:
+            return Response('already liked', status=status.HTTP_400_BAD_REQUEST)
         like.delete()
         result = 'disliked'
-    except LikeComment.DoesNotExist:
-        LikeComment.objects.create(user=request.user, comment_id=pk)
-        result = 'liked'
 
     return Response(result, status=status.HTTP_200_OK)
